fix: real time gi

+++ Shaders/ShaderLibrarys/UnityInput.hlsl
@@ -105,13 +105,13 @@ CBUFFER_START(UnityPerDraw)
     float4 unity_LightmapST;
     float4 unity_DynamicLightmapST;
 
-    half4 unity_SHAr;
-    half4 unity_SHAg;
-    half4 unity_SHAb;
-    half4 unity_SHBr;
-    half4 unity_SHBg;
-    half4 unity_SHBb;
-    half4 unity_SHC;
+    float4 unity_SHAr;
+    float4 unity_SHAg;
+    float4 unity_SHAb;
+    float4 unity_SHBr;
+    float4 unity_SHBg;
+    float4 unity_SHBb;
+    float4 unity_SHC;
 
     // Renderer bounding box
     float4 unity_RendererBounds_Min;
